tests/requirements.txt
asyncmock==0.4.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
execnet==1.9.0
grpcio==1.51.3
grpcio-tools==1.51.3
idna==3.4
iniconfig==2.0.0
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
protobuf==4.22.1
PyJWT==2.6.0
pytest==7.2.2
pytest-grpc==0.8.0
pytest-httpserver==1.0.6
python-ulid==1.1.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
